commiting changes made on 03/02/2023

+++ tenmo-server/src/main/java/com/techelevator/tenmo/dao/AccountDao.java
@@ -7,8 +7,8 @@ import java.math.BigDecimal;
 public interface AccountDao {
 
     BigDecimal findBalanceById(int userId);
-    boolean addBalance(int balanceToAdd, int userId);
-    boolean subtractBalance(int balanceToSubtract, int userId);
+    boolean addBalance(BigDecimal balanceToAdd, int userId);
+    boolean subtractBalance(BigDecimal balanceToSubtract, int userId);
     Account findAccountById(int userId);
 
 }
